+++ HPXMLtoOpenStudio/measure.rb
@@ -214,10 +214,25 @@ class OSModel
     @apply_ashrae140_assumptions = @hpxml.header.apply_ashrae140_assumptions # Hidden feature
     @apply_ashrae140_assumptions = false if @apply_ashrae140_assumptions.nil?
 
-    # Check paths
+    # Check/update file references
     @hpxml.header.schedules_filepath = FilePath.check_path(@hpxml.header.schedules_filepath,
                                                            File.dirname(hpxml_path),
                                                            'Schedules')
+    @hpxml.header.co2_emissions_scenarios.each do |scenario|
+      scenario.elec_schedule_filepath = FilePath.check_path(scenario.elec_schedule_filepath,
+                                                            File.dirname(hpxml_path),
+                                                            'CO2 Emissions File')
+      data = File.readlines(scenario.elec_schedule_filepath)
+      if data.size != 8760
+        fail "CO2 Emissions File has invalid number of rows (#{data.size}). Must be 8760."
+      end
+      if data.select { |x| x.include? ',' }.size > 0
+        fail 'CO2 Emissions File has multiple columns. Must be a single column of data.'
+      end
+      if data.map(&:strip).map { |x| Float(x) rescue nil }.any? nil
+        fail 'CO2 Emissions File has non-numeric values.'
+      end
+    end
 
     # Init
 
@@ -2075,6 +2090,8 @@ class OSModel
     additionalProperties.setFeature('hpxml_path', hpxml_path)
     additionalProperties.setFeature('hpxml_defaults_path', @hpxml_defaults_path)
     additionalProperties.setFeature('building_id', building_id.to_s)
+    co2_emissions_scenario_names = @hpxml.header.co2_emissions_scenarios.map { |s| s.name }.to_s
+    additionalProperties.setFeature('co2_emissions_scenario_names', co2_emissions_scenario_names)
   end
 
   def self.map_to_string(map)